+++ src/settings/settings.module.ts
@@ -3,11 +3,19 @@ import { SettingsController } from './settings.controller';
 import { SettingsService } from './settings.service';
 import { BookdarrModule } from '../bookdarr/bookdarr.module';
 import { AuthModule } from '../auth/auth.module';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { SmtpConfigEntity } from './smtp-config.entity';
+import { SmtpConfigService } from './smtp-config.service';
+import { SmtpConfigController } from './smtp-config.controller';
 
 @Module({
-  imports: [forwardRef(() => BookdarrModule), forwardRef(() => AuthModule)],
-  controllers: [SettingsController],
-  providers: [SettingsService],
-  exports: [SettingsService],
+  imports: [
+    forwardRef(() => BookdarrModule),
+    forwardRef(() => AuthModule),
+    TypeOrmModule.forFeature([SmtpConfigEntity]),
+  ],
+  controllers: [SettingsController, SmtpConfigController],
+  providers: [SettingsService, SmtpConfigService],
+  exports: [SettingsService, SmtpConfigService],
 })
 export class SettingsModule {}